+++ src/main/java/pl/pkrysztofiak/gridview/view/layout/grid/GridPanelsView.java
@@ -10,8 +10,8 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.layout.Pane;
 import pl.pkrysztofiak.gridview.model.layout.grid.GridPanelsModel;
-import pl.pkrysztofiak.gridview.model.panels.grid.VGridLineModel;
-import pl.pkrysztofiak.gridview.view.panels.grid.VGridLineView;
+import pl.pkrysztofiak.gridview.model.layout.grid.lines.VGridLineModel;
+import pl.pkrysztofiak.gridview.view.layout.grid.lines.VGridLineView;
 
 public class GridPanelsView extends Pane {
